Se crearon nuevos permisos

See merge request silent4business/tabantaj!699

+++ app/Http/Controllers/Admin/PerfilController.php
@@ -6,12 +6,14 @@ use App\Http\Controllers\Controller;
 use App\Models\Empleado;
 use App\Models\PerfilEmpleado;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 use Yajra\DataTables\Facades\DataTables;
 
 class PerfilController extends Controller
 {
     public function index(Request $request)
     {
+        abort_if('niveles_jerarquicos_access', Response::HTTP_FORBIDDEN, '403 Forbidden');
         if ($request->ajax()) {
             $query = PerfilEmpleado::with(['empleados'])->orderBy('id')->get();
             $table = DataTables::of($query);
@@ -20,9 +22,9 @@ class PerfilController extends Controller
             $table->addColumn('actions', '&nbsp;');
 
             $table->editColumn('actions', function ($row) {
-                $viewGate = 'user_show';
-                $editGate = 'user_edit';
-                $deleteGate = 'user_delete';
+                $viewGate = 'niveles_jerarquicos_create';
+                $editGate = 'niveles_jerarquicos_edit';
+                $deleteGate = 'niveles_jerarquicos_delete';
                 $crudRoutePart = 'perfiles';
 
                 return view('partials.datatablesActions', compact(
@@ -54,6 +56,7 @@ class PerfilController extends Controller
 
     public function create()
     {
+        abort_if('niveles_jerarquicos_create', Response::HTTP_FORBIDDEN, '403 Forbidden');
         $empleados = Empleado::get();
 
         return view('admin.perfiles.create', compact('empleados'));
@@ -61,6 +64,8 @@ class PerfilController extends Controller
 
     public function store(Request $request)
     {
+        abort_if('niveles_jerarquicos_create', Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $perfil = PerfilEmpleado::create($request->all());
 
         return redirect()->route('admin.perfiles.index')->with('success', 'Guardado con éxito');
@@ -68,6 +73,7 @@ class PerfilController extends Controller
 
     public function edit($perfil)
     {
+        abort_if('niveles_jerarquicos_edit', Response::HTTP_FORBIDDEN, '403 Forbidden');
         $perfil = PerfilEmpleado::find($perfil);
 
         return view('admin.perfiles.edit', compact('perfil'));
@@ -75,6 +81,7 @@ class PerfilController extends Controller
 
     public function update(Request $request, $perfil)
     {
+        abort_if('niveles_jerarquicos_edit', Response::HTTP_FORBIDDEN, '403 Forbidden');
         $perfil = PerfilEmpleado::find($perfil);
         $perfil->update($request->all());
 
@@ -83,6 +90,7 @@ class PerfilController extends Controller
 
     public function show($perfil)
     {
+        abort_if('niveles_jerarquicos_show', Response::HTTP_FORBIDDEN, '403 Forbidden');
         $perfil = PerfilEmpleado::find($perfil);
 
         return view('admin.perfiles.show', compact('perfil'));
@@ -90,6 +98,7 @@ class PerfilController extends Controller
 
     public function destroy($perfil)
     {
+        abort_if('niveles_jerarquicos_delete', Response::HTTP_FORBIDDEN, '403 Forbidden');
         $perfil = PerfilEmpleado::find($perfil);
         // dd($perfil);
         $perfil->delete();
